Refactor Marquee component, optimize drag handler with Pointer Events and CSS
