+++ Components/Marquee.razor.cs
@@ -1,8 +1,3 @@
-using System.Globalization;
-using System.Text;
-using Microsoft.AspNetCore.Components;
-using Microsoft.JSInterop;
-
 namespace BlazorFastMarquee;
 
 public partial class Marquee : ComponentBase, IAsyncDisposable
@@ -47,7 +42,6 @@ public partial class Marquee : ComponentBase, IAsyncDisposable
   private bool _isDisposed;
 
   private bool _onMountInvoked;
-  private CancellationTokenSource? _cts;
   private readonly SemaphoreSlim _dragHandlerLock = new(1, 1);
 
   // Layout state
@@ -87,7 +81,6 @@ public partial class Marquee : ComponentBase, IAsyncDisposable
   private bool _prevPauseOnHover;
   private bool _prevPauseOnClick;
   private bool _prevEnableDrag;
-  private bool _enableDragChanged;
   private MarqueeDirection _prevDirection;
   private double _prevSpeed;
   private double _prevDelay;
@@ -106,7 +99,6 @@ public partial class Marquee : ComponentBase, IAsyncDisposable
   public Marquee()
   {
     _dotNetRef = DotNetObjectReference.Create(this);
-    _cts = new CancellationTokenSource();
   }
 
   #endregion Constructor
@@ -265,28 +257,6 @@ public partial class Marquee : ComponentBase, IAsyncDisposable
 
   #region Lifecycle Methods
 
-  protected override bool ShouldRender()
-  {
-    // Only render if parameters actually changed
-    // This prevents cascade renders from parent components
-    var shouldRender =
-      _classInvalidated
-      || _containerStyleInvalidated
-      || _gradientStyleInvalidated
-      || _marqueeStyleInvalidated
-      || _childStyleInvalidated
-      || _enableDragChanged;
-
-    return shouldRender;
-  }
-
-  protected override void OnParametersSet()
-  {
-    // Granular invalidation - only invalidate what changed
-    InvalidateCachesForParameterChanges();
-    ProcessAdditionalAttributes();
-  }
-
   protected override async Task OnAfterRenderAsync(bool firstRender)
   {
     if (_isDisposed)
@@ -301,9 +271,11 @@ public partial class Marquee : ComponentBase, IAsyncDisposable
       await EnsureAnimationHandlerAsync();
       await EnsureDragHandlerAsync();
 
-      _enableDragChanged = false;
-
-      await MeasureAsync();
+      if (firstRender || _measureRequested || _observer is null)
+      {
+        await MeasureAsync();
+        _measureRequested = false;
+      }
 
       if (firstRender && OnMount.HasDelegate && !_onMountInvoked)
       {
@@ -551,511 +523,6 @@ public partial class Marquee : ComponentBase, IAsyncDisposable
   }
 
   #endregion JS Interop Methods
-
-  #region Style Building Methods (Zero-Allocation Optimized)
-
-  private string BuildCssClass()
-  {
-    var hasClassName = !string.IsNullOrWhiteSpace(ClassName);
-    var hasAdditional = !string.IsNullOrWhiteSpace(_additionalClass);
-
-    // Fast paths for common cases (zero allocation)
-    if (!hasClassName && !hasAdditional)
-      return BaseClassName;
-
-    if (!hasAdditional)
-      return string.Concat(BaseClassName, " ", ClassName);
-
-    if (!hasClassName)
-      return string.Concat(BaseClassName, " ", _additionalClass);
-
-    return string.Concat(BaseClassName, " ", ClassName, " ", _additionalClass);
-  }
-
-  private string BuildContainerStyle()
-  {
-    var builder = GetStringBuilder(256);
-
-    // Use CSS variable names that match React Fast Marquee
-    var pauseOnHover = (!Play || PauseOnHover) ? AnimationPaused : AnimationRunning;
-    var pauseOnClick =
-      (!Play || PauseOnClick || (PauseOnHover && !PauseOnClick))
-        ? AnimationPaused
-        : AnimationRunning;
-    var width = IsVertical(Direction) ? WidthViewport : WidthFull;
-    var transform = GetContainerTransform(Direction);
-
-    AppendCssVariable(builder, "pause-on-hover", pauseOnHover);
-    AppendCssVariable(builder, "pause-on-click", pauseOnClick);
-    AppendCssVariable(builder, "width", width);
-    AppendCssVariable(builder, "transform", transform);
-    AppendRawStyle(builder, Style);
-    AppendRawStyle(builder, _additionalStyle);
-
-    return builder.ToString();
-  }
-
-  private string BuildGradientStyle()
-  {
-    var builder = GetStringBuilder(128);
-
-    // Convert CssLength to proper CSS value
-    var gradientWidthValue = GradientWidth.ToString();
-
-    AppendCssVariable(builder, "gradient-color", GradientColor);
-    AppendCssVariable(builder, "gradient-width", gradientWidthValue);
-
-    return builder.ToString();
-  }
-
-  private string BuildMarqueeStyle()
-  {
-    var builder = GetStringBuilder(256);
-
-    // Use CSS variable names that match React Fast Marquee
-    var playState = Play ? AnimationRunning : AnimationPaused;
-    var direction =
-      (Direction is MarqueeDirection.Left or MarqueeDirection.Up)
-        ? DirectionNormal
-        : DirectionReverse;
-    var minWidth = AutoFill ? MinWidthAuto : WidthFull;
-    var iterationCount = Loop > 0 ? Loop.ToString(CultureInfo.InvariantCulture) : IterationInfinite;
-
-    AppendCssVariable(builder, "play", playState);
-    AppendCssVariable(builder, "direction", direction);
-    AppendCssVariableNumber(builder, "duration", GetDuration());
-    AppendCssVariableNumber(builder, "delay", Math.Max(0, Delay));
-    AppendCssVariable(builder, "iteration-count", iterationCount);
-    AppendCssVariable(builder, "min-width", minWidth);
-
-    return builder.ToString();
-  }
-
-  private string BuildChildStyle()
-  {
-    var builder = GetStringBuilder(64);
-    var transform = GetChildTransform(Direction);
-
-    AppendCssVariable(builder, "transform", transform);
-
-    return builder.ToString();
-  }
-
-  #endregion Style Building Methods (Zero-Allocation Optimized)
-
-  #region Layout Calculation Methods
-
-  private bool ApplyLayout(double containerSpan, double marqueeSpan)
-  {
-    var hasChanged = false;
-
-    if (!AreClose(containerSpan, _containerSpan))
-    {
-      _containerSpan = containerSpan;
-      hasChanged = true;
-    }
-
-    if (!AreClose(marqueeSpan, _marqueeSpan))
-    {
-      _marqueeSpan = marqueeSpan;
-      hasChanged = true;
-    }
-
-    var newMultiplier = CalculateMultiplier(containerSpan, marqueeSpan);
-    if (newMultiplier != _multiplier)
-    {
-      _multiplier = newMultiplier;
-      hasChanged = true;
-    }
-
-    return hasChanged;
-  }
-
-  private int CalculateMultiplier(double containerSpan, double marqueeSpan)
-  {
-    if (!AutoFill || marqueeSpan <= 0d)
-      return 1;
-
-    return Math.Max(1, (int)Math.Ceiling(containerSpan / marqueeSpan));
-  }
-
-  private double GetDuration()
-  {
-    var effectiveSpeed = Math.Max(1d, Speed);
-
-    if (_marqueeSpan <= 0d)
-      return 0d;
-
-    // Match React Fast Marquee duration calculation
-    if (AutoFill)
-    {
-      return (_marqueeSpan * _multiplier) / effectiveSpeed;
-    }
-    else
-    {
-      // If marquee is smaller than container, use container width for duration
-      // Otherwise use marquee width
-      return _marqueeSpan < _containerSpan
-        ? _containerSpan / effectiveSpeed
-        : _marqueeSpan / effectiveSpeed;
-    }
-  }
-
-  #endregion Layout Calculation Methods
-
-  #region Helper Methods
-
-  private void InvalidateCachesForParameterChanges()
-  {
-    // Granular cache invalidation based on what actually changed
-    if (_prevClassName != ClassName || _prevClassName != _additionalClass)
-    {
-      _classInvalidated = true;
-      _prevClassName = ClassName;
-    }
-
-    if (
-      _prevStyle != Style
-      || _prevPlay != Play
-      || _prevPauseOnHover != PauseOnHover
-      || _prevPauseOnClick != PauseOnClick
-      || _prevDirection != Direction
-    )
-    {
-      _containerStyleInvalidated = true;
-      _prevStyle = Style;
-      _prevPlay = Play;
-      _prevPauseOnHover = PauseOnHover;
-      _prevPauseOnClick = PauseOnClick;
-      _prevDirection = Direction;
-    }
-
-    // Track EnableDrag separately to ensure proper drag handler lifecycle
-    if (_prevEnableDrag != EnableDrag)
-    {
-      _enableDragChanged = true;
-      _containerStyleInvalidated = true;
-      _prevEnableDrag = EnableDrag;
-    }
-
-    if (
-      _prevGradient != Gradient
-      || _prevGradientColor != GradientColor
-      || !_prevGradientWidth.Equals(GradientWidth)
-    )
-    {
-      _gradientStyleInvalidated = true;
-      _prevGradient = Gradient;
-      _prevGradientColor = GradientColor;
-      _prevGradientWidth = GradientWidth;
-    }
-
-    if (
-      _prevPlay != Play
-      || _prevDirection != Direction
-      || !AreClose(_prevSpeed, Speed)
-      || !AreClose(_prevDelay, Delay)
-      || _prevLoop != Loop
-      || _prevAutoFill != AutoFill
-    )
-    {
-      _marqueeStyleInvalidated = true;
-      _prevSpeed = Speed;
-      _prevDelay = Delay;
-      _prevLoop = Loop;
-      _prevAutoFill = AutoFill;
-    }
-
-    if (_prevDirection != Direction)
-    {
-      _childStyleInvalidated = true;
-    }
-  }
-
-  private void ProcessAdditionalAttributes()
-  {
-    if (AdditionalAttributes is null)
-    {
-      _additionalClass = null;
-      _additionalStyle = null;
-      _attributesWithoutClassOrStyle = null;
-      return;
-    }
-
-    var hasClass = AdditionalAttributes.TryGetValue("class", out var classAttribute);
-    var hasStyle = AdditionalAttributes.TryGetValue("style", out var styleAttribute);
-
-    _additionalClass = hasClass ? classAttribute as string : null;
-    _additionalStyle = hasStyle ? styleAttribute as string : null;
-
-    if (!hasClass && !hasStyle)
-    {
-      _attributesWithoutClassOrStyle = AdditionalAttributes;
-      return;
-    }
-
-    // Lazy allocation: only create dictionary if we need to filter
-    // Pre-size for efficiency
-    var filtered = new Dictionary<string, object>(
-      AdditionalAttributes.Count - (hasClass ? 1 : 0) - (hasStyle ? 1 : 0)
-    );
-    foreach (var (key, value) in AdditionalAttributes)
-    {
-      if (key != "class" && key != "style")
-        filtered[key] = value;
-    }
-
-    _attributesWithoutClassOrStyle = filtered;
-  }
-
-  // Reuse StringBuilder to reduce allocations
-  private StringBuilder GetStringBuilder(int capacity)
-  {
-    if (_stringBuilder is null)
-    {
-      _stringBuilder = new StringBuilder(capacity);
-    }
-    else
-    {
-      _stringBuilder.Clear();
-      if (_stringBuilder.Capacity < capacity)
-        _stringBuilder.Capacity = capacity;
-    }
-
-    return _stringBuilder;
-  }
-
-  private static bool IsVertical(MarqueeDirection direction) =>
-    direction is MarqueeDirection.Up or MarqueeDirection.Down;
-
-  private static bool IsReversedDirection(MarqueeDirection direction) =>
-    direction is MarqueeDirection.Right or MarqueeDirection.Up;
-
-  private static bool AreClose(double left, double right) => Math.Abs(left - right) < 0.1d;
-
-  // Zero-allocation transform getters using pre-allocated strings
-  private static string GetContainerTransform(MarqueeDirection direction) =>
-    direction switch
-    {
-      MarqueeDirection.Up => RotateNeg90,
-      MarqueeDirection.Down => Rotate90,
-      _ => TransformNone
-    };
-
-  private static string GetChildTransform(MarqueeDirection direction) =>
-    direction switch
-    {
-      MarqueeDirection.Up => Rotate90,
-      MarqueeDirection.Down => RotateNeg90,
-      _ => TransformNone
-    };
-
-  // Optimized CSS variable append (no string formatting)
-  private static void AppendCssVariable(StringBuilder builder, string name, string value)
-  {
-    builder.Append("--");
-    builder.Append(name);
-    builder.Append(':');
-    builder.Append(value);
-    builder.Append(';');
-  }
-
-  // Specialized version for numeric values (reduces allocations)
-  private static void AppendCssVariableNumber(StringBuilder builder, string name, double value)
-  {
-    builder.Append("--");
-    builder.Append(name);
-    builder.Append(':');
-    builder.Append(value.ToString("0.###", CultureInfo.InvariantCulture));
-    builder.Append("s;");
-  }
-
-  private static void AppendRawStyle(StringBuilder builder, string? value)
-  {
-    if (string.IsNullOrWhiteSpace(value))
-      return;
-
-    var trimmed = value.Trim();
-    if (trimmed.Length == 0)
-      return;
-
-    if (builder.Length > 0 && builder[^1] != ';')
-      builder.Append(';');
-
-    builder.Append(trimmed);
-    if (!trimmed.EndsWith(';'))
-      builder.Append(';');
-  }
-
-  #endregion Helper Methods
-
-  #region Disposal
-
-  /// <summary>
-  /// Disposes resources and cleans up JS interop references.
-  /// Handles both WASM and Server scenarios gracefully.
-  /// </summary>
-  public async ValueTask DisposeAsync()
-  {
-    if (_isDisposed)
-      return;
-
-    _isDisposed = true;
-
-    // Cancel any pending async operations
-    _cts?.Cancel();
-
-    // Dispose .NET object reference
-    _dotNetRef?.Dispose();
-
-    // Cleanup JS interop resources
-    await DisposeDragHandlerAsync();
-    await DisposeAnimationHandlerAsync();
-    await DisposeObserverAsync();
-    await DisposeModuleAsync();
-
-    // Dispose cancellation token
-    _cts?.Dispose();
-    _cts = null;
-    
-    // Dispose semaphore
-    _dragHandlerLock?.Dispose();
-  }
-
-  private async ValueTask DisposeDragHandlerAsync()
-  {
-    if (_dragHandler is null)
-      return;
-
-    var handlerToDispose = _dragHandler;
-    _dragHandler = null; // Clear immediately to prevent re-entry
-
-    try
-    {
-      await handlerToDispose.InvokeVoidAsync("dispose");
-    }
-    catch (JSDisconnectedException)
-    {
-      // Circuit already disconnected - expected
-    }
-    catch (ObjectDisposedException)
-    {
-      // Already disposed - expected
-    }
-    catch
-    {
-      // Suppress other errors during disposal
-    }
-
-    try
-    {
-      await handlerToDispose.DisposeAsync();
-    }
-    catch
-    {
-      // Suppress disposal errors
-    }
-  }
-
-  private async ValueTask DisposeAnimationHandlerAsync()
-  {
-    if (_animationHandler is null)
-      return;
-
-    try
-    {
-      await _animationHandler.InvokeVoidAsync("dispose");
-    }
-    catch (JSDisconnectedException)
-    {
-      // Circuit already disconnected - expected
-    }
-    catch (ObjectDisposedException)
-    {
-      // Already disposed - expected
-    }
-    catch
-    {
-      // Suppress other errors during disposal
-    }
-    finally
-    {
-      try
-      {
-        await _animationHandler.DisposeAsync();
-      }
-      catch
-      {
-        // Suppress disposal errors
-      }
-      _animationHandler = null;
-    }
-  }
-
-  private async ValueTask DisposeObserverAsync()
-  {
-    if (_observer is null)
-      return;
-
-    try
-    {
-      // Try to notify JS to cleanup
-      await _observer.InvokeVoidAsync("dispose");
-    }
-    catch (JSDisconnectedException)
-    {
-      // Circuit already disconnected - expected
-    }
-    catch (ObjectDisposedException)
-    {
-      // Already disposed - expected
-    }
-    catch
-    {
-      // Suppress other errors during disposal
-    }
-    finally
-    {
-      try
-      {
-        await _observer.DisposeAsync();
-      }
-      catch
-      {
-        // Suppress disposal errors
-      }
-      _observer = null;
-    }
-  }
-
-  private async ValueTask DisposeModuleAsync()
-  {
-    if (_module is null)
-      return;
-
-    try
-    {
-      await _module.DisposeAsync();
-    }
-    catch (JSDisconnectedException)
-    {
-      // Circuit already disconnected - expected
-    }
-    catch (ObjectDisposedException)
-    {
-      // Already disposed - expected
-    }
-    catch
-    {
-      // Suppress other errors during disposal
-    }
-    finally
-    {
-      _module = null;
-    }
-  }
-
-  #endregion Disposal
 
   #region Nested Types
