+++ GlobalUsings.cs
@@ -1,0 +1,5 @@
+global using Microsoft.AspNetCore.Components;
+global using Microsoft.JSInterop;
+global using System.Globalization;
+global using System.Text;
+
